Update About-Low-Level-Terms.adoc
Removed two questions to create three issues about the text in this page: 
Issue #1. is "discovered" the best word for identification of significant shortcomings?

Issue #2. do we want to find a new name for mid-level terms?

Issue #3. is this document where we should describe how low-level terms were manually assigned to mid-level terms?